% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{mdUI}
\alias{mdUI}
\title{按日期同步界面}
\usage{
mdUI(
  tabTitle = "按日期同步",
  colTitles = c("客户操作区域", "供应商操作区域", ""),
  widthRates = c(6, 6, 12),
  func_left = mdUI_left,
  func_right = mdUI_right,
  func_bottom = mdUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按日期同步界面
}
\examples{
mdUI()
}
